Added NetVar m_flFlashDuration
I have added m_flFlashDuration to the NetVar list.  It is just 1 line, and it will allow users of the base to develop Features like NoFlash, Aimbot (if they are checking if the player is flashed), or adding Flash Information to the ESP. This would be extremely easy to add for anyone with minimal experience, but I believe the idea of this base was to make it Beginner-Friendly.  Moreover, they would still need to code their own features using it so I think it's a good addition.
You probably just forgot about it :)